Trap when an argument to an allocation function overflows
Overview:

Introduce an llvm pass, ProtectLibcalls, that causes a program to trap
instead of allocating memory if the allocation size overflowed. Apple's
security team has root-caused many issues to this type of bug.

Implementation details:

The pass recursively looks for overflow in OverflowingBinaryOperator
values. It looks past basic block boundaries and integer casts. The pass
runs late in the scalar pipeline to take advantage of inlining context.

An OverflowingBinaryOperator is checked for signed overflow when `nsw`
is present and unsigned overflow otherwise (modulo some special cases).
This avoids breaking code that traffics in negative offsets. Generally,
unsigned overflow checking is preferred: it detects when an allocation
size wraps around to a small positive value, which is important for
preventing buffer overflow.

Alternatives:

Checking can be done in the frontend (e.g. CGExprScalar in clang). It
may be more difficult to check arithmetic spread out across multiple
basic blocks, which is common (esp. due to ternary operators). It may
not be possible to check inlined arithmetic.

Checking can be enabled via existing -fsanitize=integer options. This
has a very high risk of causing crashes due to false positives, as a lot
of code in the wild relies on wrap-on-unsigned-overflow behavior. Hash
functions are perhaps the most frequently touted example, but other
idiomatic cases exist.

Testing:

`check-llvm` and `check-clang` with a stage2 build with ProtectLibcalls
runtime checking enabled. World build of iOS.

rdar://58453690